Sinks are now part of the barriers_fully_acked stage of barrier protocol
A barrier is injected at all sources, and the BarrierInitiator waits
until all sinks have acked before informing all sources that it has
been fully acked. This commit removes some ambiguity in naming around
this process and adds sinks to this final stage. This is so that sinks
that have effectively precommitted state as part of checkpointing can
now commit that state.