Add support for sinks to abort checkpoints
It is possible that a sink implementation might encounter conditions
that require an in progress checkpoint to be aborted. For example,
a precommit of its outputs might fail. In this case, the sink
can abort the relevant barrier which will lead to an abort_checkpoint
call on the CheckpointInitiator.  At this point, our only option is
to roll back the cluster to the last successful checkpoint.