Spin up SourceListeners in Phase II but start source in Phase III
In order to prevent messages from being processed on one worker before
all workers are ready, we moved the creation of SourceListeners to
after Phase III was complete. However, this undermines barrier
assumptions for downstream workers when a worker with sources recovers
from failure. We need the source actors to exist before Phase III
completes so that they can forward barriers to downstreams that expect
them. This commit separates the SourceListener creation step from
the start sources step. The former happens when Phase II is complete,
the latter happens when Phase III is complete.